+++ examples/example-ethers-v5/index.js
@@ -3,7 +3,7 @@ import './index.css';
 import { ethers } from 'ethers';
 import { createPublicClient, createWalletClient, custom, http } from 'viem';
 import * as chains from 'viem/chains';
-import HolyheldSDK, { getNetworkChainId, getNetwork } from '@holyheld/sdk';
+import HolyheldSDK from '@holyheld/sdk';
 import { getSpinnerHTML, getSettingsHTML, getRadioItemHTML, getTokenInfoHTML, getDataHTML } from './templates';
 
 const parentElement = document.querySelector('section');
@@ -47,7 +47,7 @@ connectButton.addEventListener('click', async () => {
 });
 
 // 1. Initialize SDK
-initializeButton.addEventListener('click', () => {
+initializeButton.addEventListener('click', async () => {
   initializeButton.setAttribute('hidden', '');
   parentElement.innerHTML = getSpinnerHTML();
 
@@ -60,6 +60,7 @@ initializeButton.addEventListener('click', () => {
   }
 
   sdk = new HolyheldSDK({ apiKey, logger: true });
+  await sdk.init();
 
   getSettingsButton.removeAttribute('hidden');
   parentElement.innerHTML = '';
@@ -126,6 +127,7 @@ getTokensButton.addEventListener('click', async () => {
         acc === '',
         current.address,
         current.network,
+        sdk.getNetwork(current.network).displayedName,
         current.name,
         current.balance,
         current.symbol
@@ -149,7 +151,7 @@ selectTokenButton.addEventListener('click', () => {
   parentElement.innerHTML = getTokenInfoHTML(
     selectedToken.name,
     selectedToken.address,
-    selectedToken.network,
+    sdk.getNetwork(selectedToken.network).displayedName,
     selectedToken.balance,
     selectedToken.symbol
   );
@@ -191,7 +193,7 @@ setAmountButton.addEventListener('click', async () => {
     parentElement.innerHTML = getTokenInfoHTML(
       selectedToken.name,
       selectedToken.address,
-      selectedToken.network,
+      sdk.getNetwork(selectedToken.network).displayedName,
       selectedToken.balance,
       selectedToken.symbol
     );
@@ -213,7 +215,7 @@ setAmountButton.addEventListener('click', async () => {
   parentElement.innerHTML = getDataHTML(
     selectedToken.name,
     selectedToken.address,
-    selectedToken.network,
+    sdk.getNetwork(selectedToken.network).displayedName,
     selectedToken.symbol,
     amount,
     amountInEUR,
@@ -226,7 +228,7 @@ setAmountButton.addEventListener('click', async () => {
 //    wallet interaction, e.g. sign permit and then send a transaction
 submitButton.addEventListener('click', async () => {
   const { chainId } = await provider.getNetwork();
-  const tokenNetworkId = getNetworkChainId(selectedToken.network);
+  const tokenNetworkId = sdk.getNetworkChainId(selectedToken.network);
 
   submitButton.setAttribute('hidden', '');
   parentElement.innerHTML = `
@@ -239,36 +241,46 @@ submitButton.addEventListener('click', async () => {
   // switch to the correct chain (network) in the wallet
   if (chainId !== tokenNetworkId) {
     try {
-      await provider.provider.request({
-        method: 'wallet_switchEthereumChain',
-        params: [{ chainId: ethers.utils.hexValue(tokenNetworkId) }]
-      })
-    } catch (error) {
-      if (error instanceof Object && error.code === 4902) {
-        const networkInfo = getNetwork(selectedToken.network);
-        await provider.provider.request({
-          method: 'wallet_addEthereumChain',
-          params: [
-            {
-              chainId: ethers.utils.hexValue(tokenNetworkId),
-              chainName: networkInfo.name,
-              rpcUrls: networkInfo.rpcUrl(),
-              nativeCurrency: {
-                name: networkInfo.baseAsset.name,
-                symbol: networkInfo.baseAsset.symbol,
-                decimals: networkInfo.baseAsset.decimals
-              },
-              blockExplorerUrls: [networkInfo.explorer]
-            }
-          ]
-        });
+      try {
         await provider.provider.request({
           method: 'wallet_switchEthereumChain',
           params: [{ chainId: ethers.utils.hexValue(tokenNetworkId) }]
-        });
-      } else {
-        throw error;
+        })
+      } catch (error) {
+        if (error instanceof Object && error.code === 4902) {
+          const networkInfo = sdk.getNetwork(selectedToken.network);
+          await provider.provider.request({
+            method: 'wallet_addEthereumChain',
+            params: [
+              {
+                chainId: ethers.utils.hexValue(tokenNetworkId),
+                chainName: networkInfo.name,
+                rpcUrls: networkInfo.rpcUrls,
+                nativeCurrency: {
+                  name: networkInfo.baseAsset.name,
+                  symbol: networkInfo.baseAsset.symbol,
+                  decimals: networkInfo.baseAsset.decimals
+                },
+                blockExplorerUrls: [networkInfo.explorerURL]
+              }
+            ]
+          });
+          await provider.provider.request({
+            method: 'wallet_switchEthereumChain',
+            params: [{ chainId: ethers.utils.hexValue(tokenNetworkId) }]
+          });
+        } else {
+          throw error;
+        }
       }
+    } catch (error) {
+      parentElement.removeChild(parentElement.querySelector('img'));
+      dlElement.innerHTML = `
+        ${dlElement.innerHTML}
+        <dt>Result:</dt>
+        <dd>failed</dd>
+      `;
+      throw error;
     }
   }