+++ examples/example-wagmi/index.js
@@ -1,11 +1,9 @@
 import './index.css';
 
-import { publicProvider } from '@wagmi/core/providers/public'
-import { configureChains, connect, createConfig, switchNetwork } from '@wagmi/core';
+import { connect, createConfig, http, switchChain, getAccount, getPublicClient, getWalletClient } from '@wagmi/core';
 import { arbitrum, avalanche, base, gnosis, mainnet, optimism, polygon, polygonZkEvm, zkSync } from '@wagmi/core/chains';
-import { InjectedConnector } from '@wagmi/connectors/injected';
-import { getPublicClient, getWalletClient, getNetwork } from '@wagmi/core';
-import HolyheldSDK, { getNetworkChainId } from '@holyheld/sdk';
+import { injected } from '@wagmi/connectors';
+import HolyheldSDK from '@holyheld/sdk';
 import { getSpinnerHTML, getSettingsHTML, getRadioItemHTML, getTokenInfoHTML, getDataHTML } from './templates';
 
 const parentElement = document.querySelector('section');
@@ -18,8 +16,7 @@ const selectTokenButton = document.querySelector('#select-token');
 const setAmountButton = document.querySelector('#set-amount');
 const submitButton = document.querySelector('#submit');
 
-const chains = [arbitrum, avalanche, base, gnosis, mainnet, optimism, polygon, polygonZkEvm, zkSync];
-
+let config;
 let sdk;
 let settings;
 let allTokens;
@@ -39,21 +36,32 @@ connectButton.addEventListener('click', async () => {
   connectButton.setAttribute('hidden', '');
   parentElement.innerHTML = getSpinnerHTML();
 
-  const { publicClient, webSocketPublicClient } = configureChains(
-    [mainnet, polygon, optimism, polygonZkEvm, gnosis, avalanche, arbitrum, zkSync, base],
-    [publicProvider()]
-  );
+  config = createConfig({
+    chains: [mainnet, polygon, optimism, polygonZkEvm, gnosis, avalanche, arbitrum, zkSync, base],
+    connectors: [
+      injected(),
+    ],
+    transports: {
+      [mainnet.id]: http(),
+      [polygon.id]: http(),
+      [optimism.id]: http(),
+      [polygonZkEvm.id]: http(),
+      [gnosis.id]: http(),
+      [avalanche.id]: http(),
+      [arbitrum.id]: http(),
+      [zkSync.id]: http(),
+      [base.id]: http(),
+    },
+  });
 
-  createConfig({ publicClient, webSocketPublicClient });
-
-  await connect({ connector: new InjectedConnector({ chains }) });
+  await connect( config, { connector: injected() });
 
   initializeButton.removeAttribute('hidden');
   parentElement.innerHTML = '';
 });
 
 // 1. Initialize SDK
-initializeButton.addEventListener('click', () => {
+initializeButton.addEventListener('click', async () => {
   initializeButton.setAttribute('hidden', '');
   parentElement.innerHTML = getSpinnerHTML();
 
@@ -66,6 +74,7 @@ initializeButton.addEventListener('click', () => {
   }
 
   sdk = new HolyheldSDK({ apiKey, logger: true });
+  await sdk.init();
 
   getSettingsButton.removeAttribute('hidden');
   parentElement.innerHTML = '';
@@ -121,7 +130,7 @@ getTokensButton.addEventListener('click', async () => {
   getTokensButton.setAttribute('hidden', '');
   parentElement.innerHTML = getSpinnerHTML();
 
-  const walletClient = await getWalletClient();
+  const walletClient = await getWalletClient(config);
 
   const { tokens } = await sdk.getWalletBalances(walletClient.account.address);
 
@@ -134,6 +143,7 @@ getTokensButton.addEventListener('click', async () => {
         acc === '',
         current.address,
         current.network,
+        sdk.getNetwork(current.network).displayedName,
         current.name,
         current.balance,
         current.symbol
@@ -157,7 +167,7 @@ selectTokenButton.addEventListener('click', () => {
   parentElement.innerHTML = getTokenInfoHTML(
     selectedToken.name,
     selectedToken.address,
-    selectedToken.network,
+    sdk.getNetwork(selectedToken.network).displayedName,
     selectedToken.balance,
     selectedToken.symbol
   );
@@ -199,7 +209,7 @@ setAmountButton.addEventListener('click', async () => {
     parentElement.innerHTML = getTokenInfoHTML(
       selectedToken.name,
       selectedToken.address,
-      selectedToken.network,
+      sdk.getNetwork(selectedToken.network).displayedName,
       selectedToken.balance,
       selectedToken.symbol
     );
@@ -221,7 +231,7 @@ setAmountButton.addEventListener('click', async () => {
   parentElement.innerHTML = getDataHTML(
     selectedToken.name,
     selectedToken.address,
-    selectedToken.network,
+    sdk.getNetwork(selectedToken.network).displayedName,
     selectedToken.symbol,
     amount,
     amountInEUR,
@@ -233,8 +243,8 @@ setAmountButton.addEventListener('click', async () => {
 // 7. Submit sending of token to recipient's debit card (this could require more than one
 //    wallet interaction, e.g. sign permit and then send a transaction
 submitButton.addEventListener('click', async () => {
-  const chainId = getNetwork().chain.id;
-  const tokenNetworkId = getNetworkChainId(selectedToken.network);
+  const { chainId } = getAccount(config);
+  const tokenNetworkId = sdk.getNetworkChainId(selectedToken.network);
 
   submitButton.setAttribute('hidden', '');
   parentElement.innerHTML = `
@@ -246,12 +256,22 @@ submitButton.addEventListener('click', async () => {
 
   // switch to the correct chain (network) in the wallet
   if (chainId !== tokenNetworkId) {
-    await switchNetwork({ chainId: tokenNetworkId })
+    try {
+      await switchChain(config, { chainId: tokenNetworkId });
+    } catch (error) {
+      parentElement.removeChild(parentElement.querySelector('img'));
+      dlElement.innerHTML = `
+        ${dlElement.innerHTML}
+        <dt>Result:</dt>
+        <dd>failed</dd>
+      `;
+      throw error;
+    }
   }
 
-  const publicClient = getPublicClient({ chainId: tokenNetworkId });
+  const publicClient = getPublicClient(config, { chainId: tokenNetworkId });
 
-  const walletClient = await getWalletClient({ chainId: tokenNetworkId });
+  const walletClient = await getWalletClient(config, { chainId: tokenNetworkId });
 
   try {
     await sdk.topup(